pyproject.toml
[project]
name = "uchimsa-backend"
version = "0.1.0"
description = "Exam preparation app for Central University"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "granian>=2.7",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "pydantic-settings>=2.0",
    "pydantic[email]>=2.0",
    "python-jose[cryptography]>=3.3",
    "bcrypt>=4.0",
    "sqladmin>=0.18",
    "anthropic>=0.30",
    "httpx>=0.27",
    "aiofiles>=23.0",
    "python-multipart>=0.0.9",
    "itsdangerous>=2.0",
]

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "testcontainers[postgres]>=4",
    "ruff>=0.6",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
